Version 2.2.0
git-svn-id: https://plugins.svn.wordpress.org/wp-imageflow2/trunk@1115305 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -1,4 +1,8 @@
 <?php 
+if ( ! defined( 'ABSPATH' ) ) {
+	exit; // Exit if accessed directly
+}
+
 //This plugin creates an entry in the options database. When the plugin will be deleted, this code will automatically delete the database entry from the options Wordpress table.
 delete_option('wpimageflow2_options');
